[Fullstack] Use newer openvswitch
Current version used is old and does not work on Bionic nodes. But as
Xenial kernels do not include the fix for local VXLAN tunnels
(bug/1684897), we still have to use a locally compiled version.

On Xenial nodes, the Queens UCA repository has openvswitch 2.9.0
On Bionic nodes, we have 2.9.2

So use the latest 2.9 release for fullstack testing

Change-Id: Ifb61daa1f14969a1d09379599081e96053488f9f
Closes-Bug: #1818632