Migrate fullstack job definition to zuulv3 syntax
Now, as neutron-functional job is moved to zuulv3 syntax,
the same can be easily done with fullstack job.

This patch also switches experimental neutron-fullstack-with-wsgi
job.

This patch removes also fullstack part from gate scripts
and it removes hooks for dstat and stack_base as both are not
used anywhere else.

Related-Bug: #1804844
Change-Id: Iac7583f01651410a3df3e349b2c843b109a8bb18