Fix neutron-status upgrade check CLI tool
Some time ago commit [1] add registering of service config
options in neutron-status CLI tool.
Some time later, commit [2] changed name of list with
service options to be in uppercase and it wasn't changed
in neutron.cmd.status module.
Because of that neutron-status CLI command was broken and
this patch fixes it by changing config options list to be
proper.

[1] https://review.openstack.org/#/c/636363/
[2] https://review.openstack.org/#/c/636719/

Change-Id: Iba891f72dfaba6f09e1fc99b0e86f308db02df06