Do not use subnet broadcast address in unit test
neutron.tests.unit.services.revisions.test_revision_plugin.\
    TestRevisionPlugin.test_port_ip_update_revises
occasionally fails in the gate, because it assumes that a random
IP+1 is also a valid IP. However the +1 address may be the subnet
broadcast address which is then rejected by neutron.

This change makes sure to use a free IP from the subnet instead of
the +1 address.

Change-Id: I1b27cac301cdc1acf426b0e71c037c157e93cd45
Closes-Bug: #1821948