Fix creating policy rules from subattributes.
In case of policy rule checks for rules like e.g.
"create_port:fixed_ips:subnet" couldn't be created to be
passed to policy enforcer because policy module could only
create rule checks for subattributes which are dict types.

With this patch checks for such rules can be created also for
attributes which are list of dicts, like e.g. fixed_ips in port
resource.

Change-Id: I02fffe77f57a513d2362df78885d327042bb8095
Closes-Bug: #1822105