Fix error 500 during segment range creation
When new network segment range is created for "vlan" network,
specify physical network is required.
It wasn't checked so there was ugly error 500 returned to the user
and ugly stacktrace in neutron logs.

Now it is validated and error 400 with proper description of
what is missing is returned.

Change-Id: Id3809a62f0e5d74949b34788b08af2e2aec879d8
Closes-Bug: #1823297
(cherry picked from commit 84ba64a6a6ede880fe3d399743343236946369fb)