Reorder classless static route (RFC3442) records
Rationale: if there are Subnet1 with Host1 and Subnet2 with Host2
and Host2 serves Network2 behind it, in order to reach Network2 it is
required to add Network2 to Subnet1's host_routes with Host2 as a
nexthop. In this case Neutron will offer in Subnet1 the following set
of RFC3442 routes:

Network2 -> Host2
Subnet2_cidr -> 0.0.0.0

which will lead to fail installing Network2 route on the Host1 since
at the moment of Network2's appearance there is no information about
Subnet2 route.

The proposed patch orders routes in such a way that 'connected' routes
(0.0.0.0) will be placed first and, thus, all subsequent routes will
be installed succesfully.

Change-Id: Ice70eb2090e1e99c72ae51601fefd7ec5cd8867a