Improve performance for agent network scheduling
Agent network scheduling is called on every get_active_networks_info()
call, which is done in regular inverals by the dhcp agent. The
scheduling part takes a significant amount of time due to
get_dhcp_agents_hosting_networks() being called once for each network.
To improve this we refactored the function in question so we are
able to retrieve all agent network mappings in one call.

This is joint work created together with Johannes Kulik