[DHCP] Don't resync network if same port is alredy in cache
Sometimes when port is created on dhcp agent's side, it may happend
that same port is already in network cache.
Before this patch if port with same IP address was already in cache,
resync was rescheduled because of duplicate IPs found in cache.

Now resync will be scheduled only if duplicate IP address belongs to
port with different MAC address or different id.

Change-Id: I23afbc10725f5dc78e3c63e6e505ef89ba8dc4a5
Closes-Bug: #1824802