Define dhcp_agents_per_network with min=1
The minimum value for dhcp_agents_per_network was being
inforced with an explicit check in code but we can do that
with simply defining the min value in the option itself.

This adds the minimum to the option definition and removes
the explicit check in the code. Note that as a result the
validate_post_plugin_load method needed to be removed
otherwise it fails the E1128 pylint check.

Change-Id: I103855d363c49fad4119ac850aed0a166d8c046d