Log when the placement service plugin is used wrongly
Log explicitly when an invalid plugin combination is requested in the
config.

The placement service plugin cannot be used with non-ML2 core
plugins. neutron-server already exits when misconfigured like that
(when a plugin raises an uncaught exception in its __init__())
so I didn't change anything about that. This change only adds a more
explicit log message for this misconfiguration.

Change-Id: Ida1a3188f8b3145cf270d280bf942b3950c0779c
Related-Bug: #1578989
Related-Bug: #1818479