Add Timer context manager class
This class creates a context that:
* Triggers a timeout exception if the timeout is set.
* Returns the time elapsed since the context was initialized.
* Returns the time spent in the context once it's closed.

The timeout exception can be suppressed; when the time expires, the
context finishes without rising TimerTimeout.

This class can be used to execute blocking methods and finish them if
there is no return. For example, in the followup patch [1], this class
is used in the method "neutron.agent.linux.ip_lib.ip_monitor" to control
how much time the blocking function "IPRSocket.get" is being called. By
using this strategy the method "ip_monitor" can be finished gracefully
using an external event signal in the main loop.

[1] https://review.opendev.org/#/c/660611/

Change-Id: I1f33535b201d49b875437bcc3397fcb465118064
Related-Bug: #1680183