Add guidelines for removal of OVO version downgrades
OVO objects include downgrade code, so that if a new object version is
introduced but an agent is running old code that does not understand it,
OVO can downgrade the object to something compatible with the older
version.  But this code should not be permanent.  This change codifies
guidelines for how long that code should exist and when it can be
removed.  It also adds pointers to information regarding the upgrade
strategies for several major distributions.

Change-Id: Ia0a6c3de932ff032aca59ac9b799faff7e261f2c