Revert "Pass network's dns_domain to dnsmasq conf"
The dns_domain attribute of a network is intended for use
by neutron when creating DNS records in an external DNS
system such as Designate.

By using the networks dns_domain, the configured search
path on booted instances mismatches with the generated
dns assignments for instance ports in the hosts file
for dnsmasq which creates a mismatched forward/reverse
lookup behaviour.

This reverts commit 28b90f6c14659500694f321eedbf313d83a2ae9c.

Change-Id: I26b5e8c35233718d97e58cfd5735c62d1684b4b0
Closes-Bug: 1826419
Depends-On: I145144c042b100f7e12a02a8ac7e0fbbe41e984d