Turn CIDR in query filter into proper subnet
Use netaddr.IPNetwork to convert CIDRs in a query
into proper subnets.

When creating subnets the CIDR is converted into
a proper subnet. i.e a subnet can be created with
cidr=10.0.0.11/24 in the create request. On create
the cidr is turned into a proper subnet resulting
in a subnet with cidr=10.0.0.0/24.

This change does the same to CIDRs when used as
query filters during a list subnet request, so
that the same value that was used to create the
subnet can be used to find the subnet.

Closes-Bug: #1831811
Change-Id: I8ae478a03ceedc6c3b1ae1d40081b5e5158813e6
(cherry picked from commit af773557323580c7b07bd41dd1f524e81530c939)