Release notes for dns_domain behavioural changes
As part of the fix for bug 1826419, a prior release note was
deleted and no new release note was added to document the
change in behaviour as a result of upgrading.

Restore deleted release note in the stable branches,
the new note will be added in the master branch.

Change-Id: Ic668d64c28cdc1068cb2413b09839a127bad46d3
Related-Bug: 1826419