[OVS] Network segmentation ID change not allowed if multisegments
If a network has several segments, the provider network segmentation ID
cannot be changed. This is defined in the feature spec [1].

In the case of having a multisegment network, the OVS agent RPC call
"get_network_details" will return the following value:

  {
    "network": {
      "updated_at": "2019-02-19T13:25:15Z",
      "revision_number": 5,
      "id": "ba973781-065b-4c69-9f07-a5c4f9a3b754",
      ...
      "segments": [{
        "provider:network_type": "vxlan",
        "provider:physical_network": null,
        "provider:segmentation_id": 10041
      },
      {
        "provider:network_type": "vxlan",
        "provider:physical_network": null,
        "provider:segmentation_id": 10066
      }],
      ...
}

The provider network information will be contained inside the "segments"
list, instead of being container in the parent "network" dictionary.

Closes-Bug: #1832745

[1]https://opendev.org/openstack/neutron-specs/src/branch/master/specs/stein/change-segmentation-id-vlan-provider-network.rst

Change-Id: I4fa37519bbf91e93ebd2f0e46e4d14edd40728fd