Remove some synchronization decorators from ip_lib
Removed the synchronization from the following ip_lib functions:
* create_netns
* remove_netns
* list_netns

The mentioned bug #1811515 and the corresponding patch [1], explain
the problem of Pyroute2 NetNS class and the file descriptors used
in a multithread environment. This is why, until a fix is applied to
this library, a synchronization decorator was applied to all Pyroute2
commands.

However, the listed functions do not instantiate this class. These three
methods only handle the filesystem in order to create, remove or list
the system namespaces. That means those methods are not affected by the
mentioned bug in Pyroute2 and it is safe to execute them without any
synchronization.

[1] https://review.opendev.org/#/c/631275/

Change-Id: I71521efbdaf9bb6d7cd9650f77c52bf6e02c9ced
Closes-Bug: #1833717
Related-Bug: #1811515